# specs/005-swagger-api-docs/contracts/openapi.yaml
openapi: "3.1.0"
info:
  title: "API-X"
  version: "0.1.0"
  description: |
    **API-X** adalah backend REST API untuk manajemen event dan pendaftaran peserta,
    dilengkapi dengan autentikasi berbasis JWT dan laporan statistik untuk admin.

    ## Cara Menggunakan API Ini

    Sebagian besar endpoint memerlukan autentikasi. Ikuti langkah berikut:

    1. **Daftar akun** — Panggil `POST /auth/register` dengan nama, email, dan kata sandi.
    2. **Login** — Panggil `POST /auth/login` dengan email dan kata sandi untuk mendapatkan JWT access token.
    3. **Salin token** — Dari respons login, salin nilai `access_token`.
    4. **Masukkan token** — Klik tombol **Authorize** di atas, lalu masukkan token dalam format `Bearer <token>`.
    5. **Akses endpoint** — Semua endpoint terproteksi kini dapat digunakan langsung dari halaman ini.

    ## Peran Pengguna

    | Peran | Keterangan |
    |-------|-----------|
    | **Publik** | Dapat diakses tanpa autentikasi |
    | **Memerlukan token** | Memerlukan JWT token pengguna terdaftar |
    | **Khusus admin** | Memerlukan JWT token dengan role `admin` |

    ## Format Error

    Semua error menggunakan struktur JSON yang konsisten:

    ```json
    {
      "error": {
        "code": "KODE_ERROR",
        "message": "Pesan yang dapat dibaca manusia.",
        "httpStatus": 4xx
      }
    }
    ```
servers:
  - url: "http://localhost:8000"
    description: "Lokal (development)"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token yang diperoleh dari `POST /auth/login`.

        **Cara mendapatkan token**:
        1. Panggil `POST /auth/login` dengan email dan kata sandi yang valid.
        2. Salin nilai `access_token` dari respons.
        3. Klik tombol **Authorize** di atas dan masukkan nilai token (tanpa awalan "Bearer").

        **Format header**: `Authorization: Bearer <token>`

        Token berlaku selama **60 menit** sejak diterbitkan (HS256 JWT).
  schemas:
...
    RegistrationResponse:
      type: object
      description: "Konfirmasi pendaftaran peserta ke sebuah event."
      properties:
        id:
          type: integer
          description: "ID unik pendaftaran."
          examples:
            - 101
        event_id:
          type: integer
          description: "ID event yang didaftarkan."
          examples:
            - 42
        status:
          type: string
          enum: [active, cancelled]
          description: "Status pendaftaran. `\"active\"` berarti terdaftar; `\"cancelled\"` berarti telah dibatalkan."
          examples:
            - "active"
        registered_at:
          type: string
          format: date-time
          description: "Waktu pendaftaran dilakukan."
          examples:
            - "2026-03-09T10:30:00"
    EventSummary:
      type: object
      description: "Ringkasan informasi event, disertakan dalam respons pendaftaran."
      properties:
        id:
          type: integer
          description: "ID unik event."
          examples:
            - 42
        title:
          type: string
          description: "Judul event."
          examples:
            - "Workshop Python Lanjutan"
        date:
          type: string
          format: date-time
          description: "Tanggal dan waktu pelaksanaan."
          examples:
            - "2026-06-15T09:00:00"
        registration_deadline:
          type: string
          format: date-time
          description: "Batas waktu pendaftaran."
          examples:
            - "2026-06-10T23:59:59"
        status:
          type: string
          enum: [active, deleted]
          description: "Status event."
          examples:
            - "active"
    RegistrationWithEventResponse:
      type: object
      description: "Pendaftaran beserta detail ringkasan event yang didaftarkan. Digunakan pada endpoint daftar pendaftaran milik pengguna."
      properties:
        id:
          type: integer
          description: "ID unik pendaftaran."
          examples:
            - 101
        event_id:
          type: integer
          description: "ID event yang didaftarkan."
          examples:
            - 42
        status:
          type: string
          enum: [active, cancelled]
          description: "Status pendaftaran."
          examples:
            - "active"
        registered_at:
          type: string
          format: date-time
          description: "Waktu pendaftaran dilakukan."
          examples:
            - "2026-03-09T10:30:00"
        cancelled_at:
          type: ["string", "null"]
          format: date-time
          description: "Waktu pembatalan pendaftaran. Bernilai `null` jika pendaftaran masih aktif."
          examples:
            - null
        event:
          $ref: "#/components/schemas/EventSummary"

    # ------------------------------------------------------------------ Reports
    EventStatItem:
      type: object
      description: "Statistik peserta untuk satu event aktif."
      properties:
        id:
          type: integer
          description: "ID unik event."
          examples:
            - 42
        title:
          type: string
          description: "Judul event."
          examples:
            - "Workshop Python Lanjutan"
        date:
          type: string
          format: date-time
          description: "Tanggal dan waktu pelaksanaan."
          examples:
            - "2026-06-15T09:00:00"
        quota:
          type: integer
          description: "Kapasitas maksimum peserta event."
          examples:
            - 50
        total_registered:
          type: integer
          description: "Jumlah peserta yang saat ini terdaftar dengan status `active` (tidak termasuk yang sudah dibatalkan)."
          examples:
            - 37
        remaining_quota:
          type: integer
          description: "Sisa kuota yang tersedia = `quota − total_registered`. Dapat bernilai negatif jika ada anomali data — tidak di-clamp ke nol agar admin dapat mendeteksinya."
          examples:
            - 13
    EventStatsPage:
      type: object
      description: "Halaman berisi statistik peserta per event aktif, beserta informasi pagination."
      properties:
        items:
          type: array
          description: "Daftar statistik event pada halaman ini."
          items:
            $ref: "#/components/schemas/EventStatItem"
        total:
          type: integer
          description: "Total event aktif di semua halaman."
          examples:
            - 120
        page:
          type: integer
          description: "Nomor halaman saat ini (dimulai dari 1)."
          examples:
            - 1
        size:
          type: integer
          description: "Jumlah item yang diminta per halaman."
          examples:
            - 20
        pages:
          type: integer
          description: "Total jumlah halaman yang tersedia."
          examples:
            - 6
    ReportSummaryResponse:
      type: object
      description: "Ringkasan jumlah event yang sedang aktif saat ini."
      properties:
        total_active_events:
          type: integer
          description: "Total event yang berstatus `active` DAN tanggal pelaksanaannya belum lewat pada saat query dijalankan."
          examples:
            - 42

    # ------------------------------------------------------------------ Pagination
...
    ErrorDetail:
      type: object
      description: "Detail error yang dikembalikan oleh API."
      required: [code, message, httpStatus]
      properties:
        code:
          type: string
          description: "Kode error yang unik dan dapat dibaca program. Gunakan nilai ini untuk penanganan error di sisi klien."
          examples:
            - "UNAUTHORIZED"
        message:
          type: string
          description: "Pesan error yang dapat dibaca manusia, menjelaskan apa yang terjadi."
          examples:
            - "Could not validate credentials."
        httpStatus:
          type: integer
          description: "Kode HTTP status, sama dengan status code respons."
          examples:
            - 401
    ErrorEnvelope:
      type: object
      description: "Struktur JSON standar untuk semua respons error di API-X."
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
      examples:
        unauthorized:
          summary: "401 Unauthorized"
          value:
            error:
              code: "UNAUTHORIZED"
              message: "Could not validate credentials."
              httpStatus: 401
        forbidden:
          summary: "403 Forbidden"
          value:
            error:
              code: "FORBIDDEN"
              message: "You do not have permission to perform this action."
              httpStatus: 403
        not_found:
          summary: "404 Not Found"
          value:
            error:
              code: "EVENT_NOT_FOUND"
              message: "Event not found."
              httpStatus: 404
        validation_error:
          summary: "422 Validation Error"
          value:
            error:
              code: "VALIDATION_ERROR"
              message: "value is not a valid email address"
              httpStatus: 422
...
paths:
...
  /registrations/{event_id}:
    post:
      tags: [registrations]
      summary: "Daftar ke sebuah event"
      description: |
        Mendaftarkan pengguna yang sedang login ke event yang ditentukan.

        **Akses**: Memerlukan token — sertakan JWT di header `Authorization: Bearer <token>`.

        **Aturan**:
        - Pengguna tidak boleh mendaftar ke event yang sudah mereka daftarkan (dan pendaftaran masih aktif).
        - Kuota event tidak boleh sudah penuh.
        - Batas waktu pendaftaran event belum boleh lewat.
      security:
        - BearerAuth: []
      parameters:
        - name: event_id
          in: path
          required: true
          description: "ID event yang ingin didaftarkan."
          schema:
            type: integer
            minimum: 1
          examples:
            valid:
              value: 42
      responses:
        "201":
          description: "Pendaftaran berhasil."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationResponse"
              example:
                id: 101
                event_id: 42
                status: "active"
                registered_at: "2026-03-09T10:30:00"
        "401":
          description: "Token tidak ada, tidak valid, atau sudah kedaluwarsa."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error:
                  code: "UNAUTHORIZED"
                  message: "Could not validate credentials."
                  httpStatus: 401
        "404":
          description: "Event tidak ditemukan."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error:
                  code: "EVENT_NOT_FOUND"
                  message: "Event not found."
                  httpStatus: 404
        "409":
          description: "Pendaftaran gagal karena salah satu kondisi: sudah terdaftar, kuota penuh, atau batas waktu lewat."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              examples:
                duplicate:
                  summary: "Sudah terdaftar ke event ini"
                  value:
                    error:
                      code: "DUPLICATE_REGISTRATION"
                      message: "You already have an active registration for this event."
                      httpStatus: 409
                quota_full:
                  summary: "Kuota event sudah penuh"
                  value:
                    error:
                      code: "QUOTA_FULL"
                      message: "Event quota is full."
                      httpStatus: 409
                deadline_passed:
                  summary: "Batas waktu pendaftaran sudah lewat"
                  value:
                    error:
                      code: "REGISTRATION_DEADLINE_PASSED"
                      message: "Registration deadline has passed."
                      httpStatus: 409

    delete:
      tags: [registrations]
      summary: "Batalkan pendaftaran dari sebuah event"
      description: |
        Membatalkan pendaftaran aktif pengguna yang sedang login dari event yang ditentukan.

        **Akses**: Memerlukan token — sertakan JWT di header `Authorization: Bearer <token>`.

        Mengembalikan **204 No Content** jika berhasil (tidak ada body respons).
      security:
        - BearerAuth: []
      parameters:
        - name: event_id
          in: path
          required: true
          description: "ID event yang ingin dibatalkan pendaftarannya."
          schema:
            type: integer
            minimum: 1
          examples:
            valid:
              value: 42
      responses:
        "204":
          description: "Pendaftaran berhasil dibatalkan. Tidak ada body respons."
        "401":
          description: "Token tidak ada, tidak valid, atau sudah kedaluwarsa."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error:
                  code: "UNAUTHORIZED"
                  message: "Could not validate credentials."
                  httpStatus: 401
        "404":
          description: "Tidak ada pendaftaran aktif yang dapat dibatalkan untuk event ini."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error:
                  code: "NO_ACTIVE_REGISTRATION"
                  message: "No active registration found for this event."
                  httpStatus: 404
  /registrations/me:
    get:
      tags: [registrations]
      summary: "Lihat daftar event yang sudah didaftarkan"
      description: |
        Mengembalikan semua pendaftaran milik pengguna yang sedang login, beserta
        ringkasan informasi setiap event yang didaftarkan.

        **Akses**: Memerlukan token — sertakan JWT di header `Authorization: Bearer <token>`.

        Mengembalikan semua pendaftaran termasuk yang sudah dibatalkan (status `cancelled`).
      security:
        - BearerAuth: []
      responses:
        "200":
          description: "Daftar pendaftaran berhasil diambil."
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RegistrationWithEventResponse"
              example:
                - id: 101
                  event_id: 42
                  status: "active"
                  registered_at: "2026-03-09T10:30:00"
                  cancelled_at: null
                  event:
                    id: 42
                    title: "Workshop Python Lanjutan"
                    date: "2026-06-15T09:00:00"
                    registration_deadline: "2026-06-10T23:59:59"
                    status: "active"
        "401":
          description: "Token tidak ada, tidak valid, atau sudah kedaluwarsa."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error:
                  code: "UNAUTHORIZED"
                  message: "Could not validate credentials."
                  httpStatus: 401

  # ================================================================ Admin Reports
  /admin/reports/events/stats:
    get:
      tags: [Admin Reporting]
      summary: "Statistik peserta per event aktif (khusus admin)"
      description: |
        Mengembalikan daftar event aktif beserta statistik jumlah peserta terdaftar
        dan sisa kuota untuk setiap event, secara terpaginasi.

        **Akses**: Khusus admin — memerlukan JWT token dengan role `admin`.

        **Definisi "event aktif"**: Event dengan status `active` DAN tanggal pelaksanaan
        belum lewat saat query dijalankan.

        **`total_registered`**: Hanya menghitung pendaftaran dengan status `active`;
        pendaftaran yang dibatalkan tidak dihitung.

        **`remaining_quota`**: `quota − total_registered`. Dapat bernilai negatif
        jika ada anomali data — tidak di-clamp ke nol agar admin dapat mendeteksinya.
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          description: "Nomor halaman yang diminta (dimulai dari 1)."
          schema:
            type: integer
            minimum: 1
            default: 1
          examples:
            first_page:
              value: 1
        - name: page_size
          in: query
          required: false
          description: "Jumlah event per halaman (minimal 1, maksimal 100)."
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          examples:
            default_size:
              value: 20
      responses:
        "200":
          description: "Statistik peserta berhasil diambil."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventStatsPage"
              example:
                items:
                  - id: 42
                    title: "Workshop Python Lanjutan"
                    date: "2026-06-15T09:00:00"
                    quota: 50
                    total_registered: 37
                    remaining_quota: 13
                  - id: 43
                    title: "Seminar Cloud Architecture"
                    date: "2026-07-01T13:00:00"
                    quota: 100
                    total_registered: 100
                    remaining_quota: 0
                total: 120
                page: 1
                size: 20
                pages: 6
        "401":
          description: "Token tidak ada, tidak valid, atau sudah kedaluwarsa."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error:
                  code: "UNAUTHORIZED"
                  message: "Could not validate credentials."
                  httpStatus: 401
        "403":
          description: "Token valid tapi bukan role admin."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error:
                  code: "FORBIDDEN"
                  message: "You do not have permission to perform this action."
                  httpStatus: 403
  /admin/reports/events/summary:
    get:
      tags: [Admin Reporting]
      summary: "Ringkasan jumlah event aktif (khusus admin)"
      description: |
        Mengembalikan total jumlah event yang sedang aktif saat ini sebagai angka tunggal.

        **Akses**: Khusus admin — memerlukan JWT token dengan role `admin`.

        **Definisi "event aktif"**: Event dengan status `active` DAN tanggal pelaksanaan
        belum lewat saat query dijalankan.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: "Ringkasan berhasil diambil."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportSummaryResponse"
              example:
                total_active_events: 42
        "401":
          description: "Token tidak ada, tidak valid, atau sudah kedaluwarsa."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error:
                  code: "UNAUTHORIZED"
                  message: "Could not validate credentials."
                  httpStatus: 401
        "403":
          description: "Token valid tapi bukan role admin."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error:
                  code: "FORBIDDEN"
                  message: "You do not have permission to perform this action."
                  httpStatus: 403